backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
python-multipart==0.0.17
openai==1.55.0
pydantic==2.10.1
python-dotenv==1.0.1
aiohttp==3.9.1
